trendspotter-mcp/requirements.txt
# Core
python-dotenv==1.0.1
requests==2.32.3

# AI Frameworks
torch==2.4.1
torchvision==0.18.1
torchaudio==2.3.1
tensorflow==2.16.1
transformers==4.44.2
sentence-transformers==3.1.1
huggingface-hub==0.24.6

# LangChain
langchain==0.2.14
langchain-community==0.2.10
langchain-openai==0.1.21

# Data / Forecasting
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
prophet==1.1.5
scipy==1.14.1
tweepy==4.14.0
mcp==0.1.0

# Visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.23.0
wordcloud==1.9.3
networkx==3.3

# Dashboard
streamlit==1.37.1
gradio==4.39.0

# Workflow Automation (Python SDK for n8n if extending)
n8npy==0.1.4

# Audio & Image Processing
pydub==0.25.1
librosa==0.10.2.post1
opencv-python==4.10.0.84
Pillow==10.4.0

# Utilities
tqdm==4.66.5
fastapi==0.114.0
uvicorn==0.30.6
